Rename notification package to notifications
All package names in a ggrc modules are in plural, so we should rename
notifications to match that as well.
