+++ test/integration/ggrc_workflows/notifications/test_cycle_start_failed_notifications.py
@@ -10,7 +10,7 @@ from datetime import datetime
 from mock import patch
 
 import os
-from ggrc import notification
+from ggrc import notifications
 from ggrc.models import Notification
 from ggrc_workflows.views import send_todays_digest_notifications
 from integration.ggrc_workflows.generator import WorkflowsGenerator
@@ -48,7 +48,7 @@ class TestCycleStartFailed(TestCase):
 
     Notification.__init__ = init_decorator(Notification.__init__)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_start_failed(self, mock_mail):
 
     wf_owner = "user@example.com"
@@ -61,16 +61,16 @@ class TestCycleStartFailed(TestCase):
       self.assert200(response)
 
     with freeze_time("2015-01-01 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(wf_owner, notif_data)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(wf_owner, notif_data)
       self.assertIn("cycle_starts_in", notif_data[wf_owner])
 
     with freeze_time("2015-03-05 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(wf_owner, notif_data)
       self.assertNotIn("cycle_started", notif_data[wf_owner])
       self.assertIn(wf_owner, notif_data)
@@ -78,13 +78,13 @@ class TestCycleStartFailed(TestCase):
 
       send_todays_digest_notifications()
 
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(wf_owner, notif_data)
 
   # TODO: investigate why next_cycle_start date remains the same after
   # start_recurring_cycles
 
-  # @patch("ggrc.notification.email.send_email")
+  # @patch("ggrc.notifications.email.send_email")
   # def test_start_failed_send_notifications(self, mock_mail):
 
   #   wf_owner = "user@example.com"
@@ -97,23 +97,23 @@ class TestCycleStartFailed(TestCase):
   #     self.assert200(response)
 
   #   with freeze_time("2015-01-01 13:39:20"):
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertNotIn(wf_owner, notif_data)
 
   #   with freeze_time("2015-01-29 13:39:20"):
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_starts_in", notif_data[wf_owner])
 
   #   with freeze_time("2015-02-05 13:39:20"):
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertNotIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_start_failed", notif_data[wf_owner])
 
   #     start_recurring_cycles()
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
@@ -121,10 +121,10 @@ class TestCycleStartFailed(TestCase):
 
   #     send_todays_digest_notifications()
 
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertNotIn(wf_owner, notif_data)
 
-  # @patch("ggrc.notification.email.send_email")
+  # @patch("ggrc.notifications.email.send_email")
   # def test_start_failed_send_notifications_monthly(self, mock_mail):
 
   #   wf_owner = "user@example.com"
@@ -139,14 +139,14 @@ class TestCycleStartFailed(TestCase):
 
   #     self.assert200(response)
 
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertNotIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_start_failed", notif_data[wf_owner])
 
   #     start_recurring_cycles()
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertIn(wf_owner, notif_data)
   #     self.assertIn("cycle_started", notif_data[wf_owner])
   #     self.assertIn(wf_owner, notif_data)
@@ -154,7 +154,7 @@ class TestCycleStartFailed(TestCase):
 
   #     send_todays_digest_notifications()
 
-  #     _, notif_data = notification.get_todays_notifications()
+  #     _, notif_data = notifications.get_todays_notifications()
   #     self.assertNotIn(wf_owner, notif_data)
 
   def create_test_cases(self):
@@ -166,7 +166,7 @@ class TestCycleStartFailed(TestCase):
       }
 
     self.quarterly_wf = {
-        "title": "quarterly wf forced notification",
+        "title": "quarterly wf forced notifications",
         "notify_on_change": True,
         "description": "",
         "owners": [person_dict(self.user.id)],
