+++ test/integration/ggrc_workflows/notifications/test_deleted_objects.py
@@ -10,7 +10,7 @@ from datetime import datetime
 from mock import patch
 
 import os
-from ggrc import notification
+from ggrc import notifications
 from ggrc.models import Notification, Person
 from ggrc_workflows.models import Workflow
 from integration.ggrc_workflows.generator import WorkflowsGenerator
@@ -48,7 +48,7 @@ class TestNotificationsForDeletedObjects(TestCase):
 
     Notification.__init__ = init_decorator(Notification.__init__)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_delete_activated_workflow(self, mock_mail):
 
     with freeze_time("2015-02-01 13:39:20"):
@@ -60,11 +60,11 @@ class TestNotificationsForDeletedObjects(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-01 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("cycle_starts_in", notif_data[user.email])
 
@@ -73,7 +73,7 @@ class TestNotificationsForDeletedObjects(TestCase):
       response = self.wf_generator.api.delete(workflow, workflow.id)
       self.assert200(response)
 
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       user = Person.query.get(self.user.id)
 
       self.assertNotIn(user.email, notif_data)
