+++ test/integration/ggrc_workflows/notifications/test_enable_disable_notifications.py
@@ -10,7 +10,7 @@ from datetime import datetime
 from mock import patch
 
 import os
-from ggrc import notification
+from ggrc import notifications
 from ggrc.models import NotificationConfig, Notification, Person
 from integration.ggrc_workflows.generator import WorkflowsGenerator
 from integration.ggrc.api_helper import Api
@@ -47,7 +47,7 @@ class TestEnableAndDisableNotifications(TestCase):
 
     Notification.__init__ = init_decorator(Notification.__init__)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_default_notificaitons_settings(self, mock_mail):
 
     with freeze_time("2015-02-01 13:39:20"):
@@ -59,14 +59,14 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-01 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_disabled_notifications(self, mock_mail):
 
     with freeze_time("2015-02-01 13:39:20"):
@@ -81,14 +81,14 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-01 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_enabled_notifications(self, mock_mail):
 
     with freeze_time("2015-02-01 13:39:20"):
@@ -99,17 +99,17 @@ class TestEnableAndDisableNotifications(TestCase):
 
     with freeze_time("2015-01-29 13:39:20"):
       user = Person.query.get(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
       self.object_generator.generate_notification_setting(
         self.user.id, "Email_Digest", True)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_forced_notifications(self, mock_mail):
 
     with freeze_time("2015-02-01 13:39:20"):
@@ -121,17 +121,17 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
       self.object_generator.generate_notification_setting(
         self.user.id, "Email_Digest", True)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_force_one_wf_notifications(self, mock_mail):
 
     with freeze_time("2015-02-01 13:39:20"):
@@ -145,7 +145,7 @@ class TestEnableAndDisableNotifications(TestCase):
       user = Person.query.get(self.user.id)
 
     with freeze_time("2015-01-29 13:39:20"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("cycle_starts_in", notif_data[user.email])
       self.assertIn(wf_forced.id, notif_data[user.email]["cycle_starts_in"])
@@ -155,7 +155,7 @@ class TestEnableAndDisableNotifications(TestCase):
         self.user.id, "Email_Digest", False)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("cycle_starts_in", notif_data[user.email])
       self.assertIn(wf_forced.id, notif_data[user.email]["cycle_starts_in"])
@@ -170,7 +170,7 @@ class TestEnableAndDisableNotifications(TestCase):
       }
 
     self.quarterly_wf_forced = {
-        "title": "quarterly wf forced notification",
+        "title": "quarterly wf forced notifications",
         "notify_on_change": True,
         "description": "",
         "owners": [person_dict(self.user.id)],
