+++ test/integration/ggrc_workflows/notifications/test_one_time_wf.py
@@ -9,7 +9,7 @@ from freezegun import freeze_time
 from mock import patch
 
 from ggrc import db
-from ggrc import notification
+from ggrc import notifications
 from ggrc.models import Notification
 from ggrc.models import Person
 from ggrc_workflows.models import Cycle
@@ -60,27 +60,27 @@ class TestOneTimeWorkflowNotification(TestCase):
       person_1 = get_person(self.random_people[0].id)
 
     with freeze_time("2015-04-11"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn("cycle_started", notif_data[person_1.email])
       self.assertIn(cycle.id, notif_data[person_1.email]["cycle_started"])
       self.assertIn("my_tasks",
                     notif_data[person_1.email]["cycle_started"][cycle.id])
 
     with freeze_time("2015-05-03"):  # two days befor due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(person_1.email, notif_data)
       self.assertNotIn("due_in", notif_data[person_1.email])
       self.assertNotIn("due_today", notif_data[person_1.email])
 
     with freeze_time("2015-05-04"):  # one day befor due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(len(notif_data[person_1.email]["due_in"]), 2)
 
     with freeze_time("2015-05-05"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(len(notif_data[person_1.email]["due_today"]), 2)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_one_time_wf_activate_single_person(self, mock_mail):
 
     with freeze_time("2015-04-10"):
@@ -92,7 +92,7 @@ class TestOneTimeWorkflowNotification(TestCase):
       self.wf_generator.activate_workflow(wf)
 
     with freeze_time("2015-04-11"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn("cycle_started", notif_data[user])
       self.assertIn(cycle.id, notif_data[user]["cycle_started"])
       self.assertIn("my_tasks", notif_data[user]["cycle_started"][cycle.id])
@@ -111,17 +111,17 @@ class TestOneTimeWorkflowNotification(TestCase):
         self.assertIn("cycle_task_url", cycle_data["cycle_tasks"][task.id])
 
     with freeze_time("2015-05-03"):  # two days befor due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user, notif_data)
       self.assertNotIn("due_in", notif_data[user])
       self.assertNotIn("due_today", notif_data[user])
 
     with freeze_time("2015-05-04"):  # one day befor due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(len(notif_data[user]["due_in"]), 3)
 
     with freeze_time("2015-05-05"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(len(notif_data[user]["due_today"]), 3)
 
       send_todays_digest_notifications()
