+++ test/integration/ggrc_workflows/notifications/test_one_time_wf_decline_accept_tasks.py
@@ -11,7 +11,7 @@ from mock import patch
 from sqlalchemy import and_
 
 import os
-from ggrc import db, notification
+from ggrc import db, notifications
 from ggrc.models import NotificationType, Notification, Person
 from ggrc_workflows.views import send_todays_digest_notifications
 from ggrc_workflows.models import Cycle, CycleTaskGroupObjectTask
@@ -134,7 +134,7 @@ class TestCycleTaskStatusChange(TestCase):
       )).all()
 
       # there is still one task in the cycle, so there should be no
-      # notification for all tasks completed
+      # notifications for all tasks completed
       self.assertEqual(notif, [])
 
       notif = db.session.query(Notification).filter(and_(
@@ -166,7 +166,7 @@ class TestCycleTaskStatusChange(TestCase):
 
       self.assertEqual(len(notif), 1, "notifications: {}".format(str(notif)))
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_single_task_declined(self, mock_mail):
     """
     test moving the end date to the future, befor due_in and due_today
@@ -188,7 +188,7 @@ class TestCycleTaskStatusChange(TestCase):
 
       self.task_change_status(task1, "Finished")
 
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-02"):
@@ -201,12 +201,12 @@ class TestCycleTaskStatusChange(TestCase):
       self.task_change_status(task1, "Declined")
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
 
       self.assertIn(user.email, notif_data)
       self.assertIn("task_declined", notif_data[user.email])
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_single_task_accepted(self, mock_mail):
     """
     test moving the end date to the future, befor due_in and due_today
@@ -228,7 +228,7 @@ class TestCycleTaskStatusChange(TestCase):
 
       self.task_change_status(task1, "Finished")
 
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-03"):
@@ -239,11 +239,11 @@ class TestCycleTaskStatusChange(TestCase):
       self.task_change_status(task1)
 
       user = Person.query.get(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
       self.assertIn("all_tasks_completed", notif_data["user@example.com"])
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_end_cycle(self, mock_mail):
     """
     manaually ending a cycle should stop all notifications for that cycle
@@ -255,7 +255,7 @@ class TestCycleTaskStatusChange(TestCase):
       self.wf_generator.activate_workflow(wf)
 
     with freeze_time("2015-05-03"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       cycle = Cycle.query.get(cycle.id)
       user = Person.query.get(self.user.id)
       self.assertIn(user.email, notif_data)
@@ -263,7 +263,7 @@ class TestCycleTaskStatusChange(TestCase):
       cycle = Cycle.query.get(cycle.id)
       self.assertFalse(cycle.is_current)
 
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertNotIn(user.email, notif_data)
 
   def create_test_cases(self):
