+++ test/integration/ggrc_workflows/notifications/test_one_time_wf_end_date_change.py
@@ -10,7 +10,7 @@ from datetime import date, datetime
 from mock import patch
 
 import os
-from ggrc import db, notification
+from ggrc import db, notifications
 from ggrc.models import Notification, Person
 from ggrc_workflows.views import send_todays_digest_notifications
 from ggrc_workflows.models import Cycle, CycleTaskGroupObjectTask
@@ -49,7 +49,7 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     Notification.__init__ = init_decorator(Notification.__init__)
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_no_date_change(self, mock_mail):
     def get_person(person_id):
       return db.session.query(Person).filter(Person.id == person_id).one()
@@ -62,11 +62,11 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-04-11 03:21:34"):
       user = get_person(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn("cycle_started", notif_data[user.email])
 
     with freeze_time("2015-05-02 03:21:34"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("cycle_started", notif_data[user.email])
       self.assertNotIn("due_in", notif_data[user.email])
@@ -74,14 +74,14 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-05-02 03:21:34"):
       send_todays_digest_notifications()
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
       # one email to owner and one to assigne
       self.assertEqual(mock_mail.call_count, 2)
 
     with freeze_time("2015-05-04 03:21:34"):  # one day befor due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       user = get_person(self.user.id)
       self.assertIn("due_in", notif_data[user.email])
       self.assertEqual(len(notif_data[user.email]["due_in"]),
@@ -89,19 +89,19 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-05-04 03:21:34"):  # one day befor due date
       send_todays_digest_notifications()
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
       # one email to owner and one to assigne
       self.assertEqual(mock_mail.call_count, 3)
 
     with freeze_time("2015-05-05 03:21:34"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn("due_today", notif_data[user.email])
       self.assertEqual(len(notif_data[user.email]["due_today"]),
                        len(self.random_objects))
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_move_end_date_to_future(self, mock_mail):
     """
     test moving the end date to the future, befor due_in and due_today
@@ -118,11 +118,11 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-04-11 03:21:34"):
       user = get_person(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn("cycle_started", notif_data[user.email])
 
     with freeze_time("2015-05-02 03:21:34"):
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("cycle_started", notif_data[user.email])
       self.assertNotIn("due_in", notif_data[user.email])
@@ -130,7 +130,7 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-05-02 03:21:34"):
       send_todays_digest_notifications()
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
       # one email to owner and one to assigne
@@ -149,23 +149,23 @@ class TestOneTimeWfEndDateChange(TestCase):
         task2, data = {"end_date": date(2015, 5, 15)})
 
     with freeze_time("2015-05-04 03:21:34"):  # one day befor due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       user = get_person(self.user.id)
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-05 03:21:34"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-14 03:21:34"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("due_in", notif_data[user.email])
       self.assertEqual(len(notif_data[user.email]["due_in"]),
                        len(self.random_objects))
 
     with freeze_time("2015-05-15 03:21:34"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
 
       # yesterdays mail has not been sent
@@ -175,7 +175,7 @@ class TestOneTimeWfEndDateChange(TestCase):
       self.assertEqual(len(notif_data[user.email]["due_today"]),
                        len(self.random_objects))
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_move_end_date_to_past(self, mock_mail):
     def get_person(person_id):
       return db.session.query(Person).filter(Person.id == person_id).one()
@@ -188,7 +188,7 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-05-02 03:21:34"):
       send_todays_digest_notifications()
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
       # one email to owner and one to assigne
@@ -207,18 +207,18 @@ class TestOneTimeWfEndDateChange(TestCase):
         task2, data = {"end_date": date(2015, 5, 1)})
 
     with freeze_time("2015-05-03 03:21:34"):  # one day befor due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-04 03:21:34"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
     with freeze_time("2015-05-05 03:21:34"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
-  @patch("ggrc.notification.email.send_email")
+  @patch("ggrc.notifications.email.send_email")
   def test_move_end_date_to_today(self, mock_mail):
     def get_person(person_id):
       return db.session.query(Person).filter(Person.id == person_id).one()
@@ -231,7 +231,7 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-05-02 03:21:34"):
       send_todays_digest_notifications()
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
       # one email to owner and one to assigne
@@ -251,7 +251,7 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-05-03 03:21:34"):  # one day befor due date
       user = get_person(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
 
       self.assertNotEquals(notif_data, {})
       self.assertIn(user.email, notif_data)
@@ -263,14 +263,14 @@ class TestOneTimeWfEndDateChange(TestCase):
 
     with freeze_time("2015-05-04 03:21:34"):  # due date
       user = get_person(self.user.id)
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertIn(user.email, notif_data)
       self.assertIn("due_today", notif_data[user.email])
       self.assertNotIn("due_in", notif_data[user.email])
       send_todays_digest_notifications()
 
     with freeze_time("2015-05-05 03:21:34"):  # due date
-      _, notif_data = notification.get_todays_notifications()
+      _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
   def create_test_cases(self):
